pami: fix build errors from C -> C++ conversion

+++ src/arch/pami/machine.C
@@ -466,7 +466,7 @@ static void short_pkt_dispatch (pami_context_t       context,
   char *smsg = (char *)pipe_addr;
   char *msg  = (char *)buffer;
 
-  CMI_CHECK_CHECKSUM(smsg, pipe_size);  
+  CMI_CHECK_CHECKSUM(smsg, (int)pipe_size);
   if (CMI_MAGIC(smsg) != CHARM_MAGIC_NUMBER) {
     /* received a non-charm msg */
     CmiAbort("Charm++ Warning: Non Charm++ Message Received. If your application has a large number of messages, this may be because of overflow in the low-level FIFOs. Please set the environment variable MUSPI_INJFIFOSIZE if the application has large number of small messages (<=4K bytes), and/or PAMI_RGETINJFIFOSIZE if the application has a large number of large messages. The default value of these variable is 65536 which is sufficient for 1000 messages in flight; please try a larger value. Please note that the memory used for these FIFOs eats up the memory = 10*FIFO_SIZE per core. Please contact Charm++ developers for further information. \n");     
@@ -722,7 +722,7 @@ void ConverseInit(int argc, char **argv, CmiStartFn fn, int usched, int initret)
 
     rc = PAMI_Context_createv (cmi_pami_client, config, cfgval, cmi_pami_contexts, _n);
     if (rc != PAMI_SUCCESS) {
-      fprintf(stderr, "PAMI_Context_createv failed for %d contexts\n", _n);
+      fprintf(stderr, "PAMI_Context_createv failed for %zu contexts\n", _n);
       assert(0);
     }
     cmi_pami_numcontexts = _n;
